+++ src/GRA.Controllers/MissionControl/ChallengesController.cs
@@ -290,6 +290,8 @@ namespace GRA.Controllers.MissionControl
         [HttpPost]
         public async Task<IActionResult> Create(ChallengesDetailViewModel model)
         {
+            byte[] badgeBytes = null;
+
             var site = await GetCurrentSiteAsync();
             if (site.MaxPointsPerChallengeTask.HasValue && model.Challenge.TasksToComplete.HasValue
                 && model.Challenge.TasksToComplete != 0)
@@ -297,15 +299,31 @@ namespace GRA.Controllers.MissionControl
                 double pointsPerChallenge = (double)model.Challenge.PointsAwarded / model.Challenge.TasksToComplete.Value;
                 if (pointsPerChallenge > site.MaxPointsPerChallengeTask)
                 {
-                    ModelState.AddModelError("Challenge.PointsAwarded", $"Too many points awarded.");
+                    ModelState.AddModelError("Challenge.PointsAwarded", "Too many points awarded.");
                 }
             }
             if (model.BadgeUploadImage != null
                 && (string.IsNullOrWhiteSpace(model.BadgeMakerImage) || !model.UseBadgeMaker))
             {
-                if (!ValidImageExtensions.Contains(Path.GetExtension(model.BadgeUploadImage.FileName).ToLower()))
+                if (!ValidImageExtensions.Contains(Path
+                    .GetExtension(model.BadgeUploadImage.FileName)
+                    .ToLowerInvariant()))
                 {
                     ModelState.AddModelError("BadgeUploadImage", $"Image must be one of the following types: {string.Join(", ", ValidImageExtensions)}");
+                }
+
+                try
+                {
+                    using (var ms = new MemoryStream())
+                    {
+                        await model.BadgeUploadImage.CopyToAsync(ms);
+                        badgeBytes = ms.ToArray();
+                    }
+                    await _badgeService.ValidateBadgeImageAsync(badgeBytes);
+                }
+                catch (GraException gex)
+                {
+                    ModelState.AddModelError("BadgeUploadImage", gex.Message);
                 }
             }
             if (ModelState.IsValid)
@@ -316,10 +334,9 @@ namespace GRA.Controllers.MissionControl
                     if (model.BadgeUploadImage != null
                         || !string.IsNullOrWhiteSpace(model.BadgeMakerImage))
                     {
-                        byte[] badgeBytes;
                         string filename;
                         if (!string.IsNullOrWhiteSpace(model.BadgeMakerImage)
-                            && (model.BadgeUploadImage != null || model.UseBadgeMaker))
+                            && (model.BadgeUploadImage == null || model.UseBadgeMaker))
                         {
                             var badgeString = model.BadgeMakerImage.Split(',').Last();
                             badgeBytes = Convert.FromBase64String(badgeString);
@@ -327,11 +344,11 @@ namespace GRA.Controllers.MissionControl
                         }
                         else
                         {
-                            using (var fileStream = model.BadgeUploadImage.OpenReadStream())
+                            if (badgeBytes == null)
                             {
                                 using (var ms = new MemoryStream())
                                 {
-                                    fileStream.CopyTo(ms);
+                                    await model.BadgeUploadImage.CopyToAsync(ms);
                                     badgeBytes = ms.ToArray();
                                 }
                             }
@@ -369,7 +386,7 @@ namespace GRA.Controllers.MissionControl
         {
             var site = await GetCurrentSiteAsync();
             var siteUrl = await _siteService.GetBaseUrl(Request.Scheme, Request.Host.Value);
-            var challenge = new Challenge();
+            Challenge challenge;
             try
             {
                 challenge = await _challengeService.MCGetChallengeDetailsAsync(id);
@@ -479,6 +496,8 @@ namespace GRA.Controllers.MissionControl
         [HttpPost]
         public async Task<IActionResult> Edit(ChallengesDetailViewModel model, string Submit)
         {
+            byte[] badgeBytes = null;
+
             var site = await GetCurrentSiteAsync();
             if (site.MaxPointsPerChallengeTask.HasValue && model.Challenge.TasksToComplete.HasValue
                 && model.Challenge.TasksToComplete != 0)
@@ -486,15 +505,31 @@ namespace GRA.Controllers.MissionControl
                 double pointsPerChallenge = (double)model.Challenge.PointsAwarded / model.Challenge.TasksToComplete.Value;
                 if (pointsPerChallenge > site.MaxPointsPerChallengeTask)
                 {
-                    ModelState.AddModelError("Challenge.PointsAwarded", $"Too many points awarded.");
+                    ModelState.AddModelError("Challenge.PointsAwarded", "Too many points awarded.");
                 }
             }
             if (model.BadgeUploadImage != null
                 && (string.IsNullOrWhiteSpace(model.BadgeMakerImage) || !model.UseBadgeMaker))
             {
-                if (!ValidImageExtensions.Contains(Path.GetExtension(model.BadgeUploadImage.FileName).ToLower()))
+                if (!ValidImageExtensions.Contains(Path
+                    .GetExtension(model.BadgeUploadImage.FileName)
+                    .ToLowerInvariant()))
                 {
                     ModelState.AddModelError("BadgeUploadImage", $"Image must be one of the following types: {string.Join(", ", ValidImageExtensions)}");
+                }
+
+                try
+                {
+                    using (var ms = new MemoryStream())
+                    {
+                        await model.BadgeUploadImage.CopyToAsync(ms);
+                        badgeBytes = ms.ToArray();
+                    }
+                    await _badgeService.ValidateBadgeImageAsync(badgeBytes);
+                }
+                catch (GraException gex)
+                {
+                    ModelState.AddModelError("BadgeUploadImage", gex.Message);
                 }
             }
             if (ModelState.IsValid)
@@ -503,10 +538,9 @@ namespace GRA.Controllers.MissionControl
                 if (model.BadgeUploadImage != null
                         || !string.IsNullOrWhiteSpace(model.BadgeMakerImage))
                 {
-                    byte[] badgeBytes;
                     string filename;
                     if (!string.IsNullOrWhiteSpace(model.BadgeMakerImage)
-                        && (model.BadgeUploadImage != null || model.UseBadgeMaker))
+                        && (model.BadgeUploadImage == null || model.UseBadgeMaker))
                     {
                         var badgeString = model.BadgeMakerImage.Split(',').Last();
                         badgeBytes = Convert.FromBase64String(badgeString);
@@ -514,11 +548,11 @@ namespace GRA.Controllers.MissionControl
                     }
                     else
                     {
-                        using (var fileStream = model.BadgeUploadImage.OpenReadStream())
+                        if (badgeBytes == null)
                         {
                             using (var ms = new MemoryStream())
                             {
-                                fileStream.CopyTo(ms);
+                                await model.BadgeUploadImage.CopyToAsync(ms);
                                 badgeBytes = ms.ToArray();
                             }
                         }
@@ -538,7 +572,9 @@ namespace GRA.Controllers.MissionControl
                         var existing = await _badgeService
                                     .GetByIdAsync((int)challenge.BadgeId);
                         existing.Filename = Path.GetFileName(model.BadgePath);
-                        await _badgeService.ReplaceBadgeFileAsync(existing, badgeBytes);
+                        await _badgeService.ReplaceBadgeFileAsync(existing,
+                            badgeBytes,
+                            model.BadgeUploadImage.FileName);
                     }
                 }
                 try
@@ -550,16 +586,17 @@ namespace GRA.Controllers.MissionControl
                         ShowAlertWarning(serviceResult.Message);
                     }
                     AlertSuccess = $"Challenge '<strong>{challenge.Name}</strong>' was successfully modified";
+
+                    var hasPermissions = UserHasPermission(Permission.ActivateAllChallenges)
+                            || UserHasPermission(Permission.ActivateSystemChallenges);
+
                     if (Submit == "Activate"
-                        && (UserHasPermission(Permission.ActivateAllChallenges)
-                            || (UserHasPermission(Permission.ActivateSystemChallenges)
-                                && challenge.RelatedSystemId == GetId(ClaimType.SystemId))))
+                        && hasPermissions
+                        && challenge.RelatedSystemId == GetId(ClaimType.SystemId)
+                        && serviceResult.Data.IsValid)
                     {
-                        if (serviceResult.Data.IsValid)
-                        {
-                            await _challengeService.ActivateChallengeAsync(serviceResult.Data);
-                            AlertSuccess = $"Challenge '<strong>{challenge.Name}</strong>' was successfully modified and activated";
-                        }
+                        await _challengeService.ActivateChallengeAsync(serviceResult.Data);
+                        AlertSuccess = $"Challenge '<strong>{challenge.Name}</strong>' was successfully modified and activated";
                     }
                 }
                 catch (GraException gex)
@@ -669,17 +706,19 @@ namespace GRA.Controllers.MissionControl
                 }
             }
 
-            foreach (string key in ModelState.Keys.Where(m => m.StartsWith("Challenge.")).ToList())
+            foreach (string key in ModelState
+                .Keys
+                .Where(m => m.StartsWith("Challenge.", StringComparison.OrdinalIgnoreCase))
+                .ToList())
             {
                 ModelState.Remove(key);
             }
 
-            if (viewModel.TaskUploadFile != null)
+            if (viewModel.TaskUploadFile != null && !ValidUploadExtensions.Contains(Path
+                    .GetExtension(viewModel.TaskUploadFile.FileName)
+                    .ToLowerInvariant()))
             {
-                if (!ValidUploadExtensions.Contains(Path.GetExtension(viewModel.TaskUploadFile.FileName).ToLower()))
-                {
-                    ModelState.AddModelError("BadgeUploadImage", $"File upload must be one of the following types: {string.Join(", ", ValidUploadExtensions)}");
-                }
+                ModelState.AddModelError("BadgeUploadImage", $"File upload must be one of the following types: {string.Join(", ", ValidUploadExtensions)}");
             }
 
             if (ModelState.IsValid)
@@ -744,17 +783,20 @@ namespace GRA.Controllers.MissionControl
                 }
             }
 
-            foreach (string key in ModelState.Keys.Where(m => m.StartsWith("Challenge.")).ToList())
+            foreach (string key in ModelState
+                .Keys
+                .Where(m => m.StartsWith("Challenge.", StringComparison.OrdinalIgnoreCase))
+                .ToList())
             {
                 ModelState.Remove(key);
             }
 
-            if (viewModel.TaskUploadFile != null)
+            if (viewModel.TaskUploadFile != null
+                && !ValidUploadExtensions.Contains(Path
+                    .GetExtension(viewModel.TaskUploadFile.FileName)
+                    .ToLowerInvariant()))
             {
-                if (!ValidUploadExtensions.Contains(Path.GetExtension(viewModel.TaskUploadFile.FileName).ToLower()))
-                {
-                    ModelState.AddModelError("BadgeUploadImage", $"File upload must be one of the following types: {string.Join(", ", ValidUploadExtensions)}");
-                }
+                ModelState.AddModelError("BadgeUploadImage", $"File upload must be one of the following types: {string.Join(", ", ValidUploadExtensions)}");
             }
 
             if (ModelState.IsValid)