+++ src/GRA.Controllers/MissionControl/TriggersController.cs
@@ -109,8 +109,6 @@ namespace GRA.Controllers.MissionControl
                     });
             }
 
-            var requireSecretCode = await GetSiteSettingBoolAsync(
-                    SiteSettingKey.Events.RequireBadge);
             foreach (var trigger in triggerList.Data)
             {
                 trigger.AwardBadgeFilename =
@@ -225,6 +223,8 @@ namespace GRA.Controllers.MissionControl
         [HttpPost]
         public async Task<IActionResult> Create(TriggersDetailViewModel model)
         {
+            byte[] badgeBytes = null;
+
             var badgeRequiredList = new List<int>();
             var challengeRequiredList = new List<int>();
             if (!string.IsNullOrWhiteSpace(model.BadgeRequiredList))
@@ -250,11 +250,29 @@ namespace GRA.Controllers.MissionControl
                 ModelState.AddModelError("BadgePath", "A badge is required.");
             }
             else if (model.BadgeUploadImage != null
-                && (string.IsNullOrWhiteSpace(model.BadgeMakerImage) || !model.UseBadgeMaker)
-                && (!ValidImageExtensions.Contains(Path.GetExtension(model.BadgeUploadImage.FileName).ToLower())))
+                    && (string.IsNullOrWhiteSpace(model.BadgeMakerImage) || !model.UseBadgeMaker))
             {
-                ModelState.AddModelError("BadgeUploadImage", $"Image must be one of the following types: {string.Join(", ", ValidImageExtensions)}");
+                if (!ValidImageExtensions.Contains(
+                    Path.GetExtension(model.BadgeUploadImage.FileName).ToLowerInvariant()))
+                {
+                    ModelState.AddModelError("BadgeUploadImage", $"Image must be one of the following types: {string.Join(", ", ValidImageExtensions)}");
+                }
+
+                try
+                {
+                    using (var ms = new MemoryStream())
+                    {
+                        await model.BadgeUploadImage.CopyToAsync(ms);
+                        badgeBytes = ms.ToArray();
+                    }
+                    await _badgeService.ValidateBadgeImageAsync(badgeBytes);
+                }
+                catch (GraException gex)
+                {
+                    ModelState.AddModelError("BadgeUploadImage", gex.Message);
+                }
             }
+
             if (!model.IsSecretCode)
             {
                 if ((!model.Trigger.Points.HasValue || model.Trigger.Points < 1)
@@ -281,12 +299,9 @@ namespace GRA.Controllers.MissionControl
                 ModelState.AddModelError("Trigger.SecretCode", "That Secret Code already exists.");
             }
 
-            if (model.AwardsPrize)
+            if (model.AwardsPrize && string.IsNullOrWhiteSpace(model.Trigger.AwardPrizeName))
             {
-                if (string.IsNullOrWhiteSpace(model.Trigger.AwardPrizeName))
-                {
-                    ModelState.AddModelError("Trigger.AwardPrizeName", "The Prize Name field is required.");
-                }
+                ModelState.AddModelError("Trigger.AwardPrizeName", "The Prize Name field is required.");
             }
             if (model.AwardsMail)
             {
@@ -311,7 +326,10 @@ namespace GRA.Controllers.MissionControl
                         model.Trigger.LimitToSystemId = null;
                         model.Trigger.LimitToBranchId = null;
                         model.Trigger.LimitToProgramId = null;
-                        model.Trigger.SecretCode = model.Trigger.SecretCode.Trim().ToLower();
+                        model.Trigger.SecretCode = model
+                            .Trigger.SecretCode
+                            .Trim()
+                            .ToLowerInvariant();
                         model.Trigger.BadgeIds = new List<int>();
                         model.Trigger.ChallengeIds = new List<int>();
                     }
@@ -335,7 +353,6 @@ namespace GRA.Controllers.MissionControl
                     if (model.BadgeUploadImage != null
                         || !string.IsNullOrWhiteSpace(model.BadgeMakerImage))
                     {
-                        byte[] badgeBytes;
                         string filename;
                         if (!string.IsNullOrWhiteSpace(model.BadgeMakerImage)
                             && (model.BadgeUploadImage == null || model.UseBadgeMaker))
@@ -346,11 +363,11 @@ namespace GRA.Controllers.MissionControl
                         }
                         else
                         {
-                            using (var fileStream = model.BadgeUploadImage.OpenReadStream())
+                            if (badgeBytes == null)
                             {
                                 using (var ms = new MemoryStream())
                                 {
-                                    fileStream.CopyTo(ms);
+                                    await model.BadgeUploadImage.CopyToAsync(ms);
                                     badgeBytes = ms.ToArray();
                                 }
                             }
@@ -492,6 +509,8 @@ namespace GRA.Controllers.MissionControl
         [HttpPost]
         public async Task<IActionResult> Edit(TriggersDetailViewModel model)
         {
+            byte[] badgeBytes = null;
+
             var badgeRequiredList = new List<int>();
             var challengeRequiredList = new List<int>();
             if (!string.IsNullOrWhiteSpace(model.BadgeRequiredList))
@@ -513,10 +532,27 @@ namespace GRA.Controllers.MissionControl
             var requirementCount = badgeRequiredList.Count + challengeRequiredList.Count;
 
             if (model.BadgeUploadImage != null
-                && (string.IsNullOrWhiteSpace(model.BadgeMakerImage) || !model.UseBadgeMaker)
-                && (!ValidImageExtensions.Contains(Path.GetExtension(model.BadgeUploadImage.FileName).ToLower())))
+                    && (string.IsNullOrWhiteSpace(model.BadgeMakerImage) || !model.UseBadgeMaker))
             {
-                ModelState.AddModelError("BadgeImage", $"Image must be one of the following types: {string.Join(", ", ValidImageExtensions)}");
+                if (!ValidImageExtensions.Contains(
+                    Path.GetExtension(model.BadgeUploadImage.FileName).ToLowerInvariant()))
+                {
+                    ModelState.AddModelError("BadgeUploadImage", $"Image must be one of the following types: {string.Join(", ", ValidImageExtensions)}");
+                }
+
+                try
+                {
+                    using (var ms = new MemoryStream())
+                    {
+                        await model.BadgeUploadImage.CopyToAsync(ms);
+                        badgeBytes = ms.ToArray();
+                    }
+                    await _badgeService.ValidateBadgeImageAsync(badgeBytes);
+                }
+                catch (GraException gex)
+                {
+                    ModelState.AddModelError("BadgeUploadImage", gex.Message);
+                }
             }
 
             if (!model.IsSecretCode)
@@ -544,12 +580,9 @@ namespace GRA.Controllers.MissionControl
             {
                 ModelState.AddModelError("Trigger.SecretCode", "That Secret Code already exists.");
             }
-            if (model.AwardsPrize)
+            if (model.AwardsPrize && string.IsNullOrWhiteSpace(model.Trigger.AwardPrizeName))
             {
-                if (string.IsNullOrWhiteSpace(model.Trigger.AwardPrizeName))
-                {
-                    ModelState.AddModelError("Trigger.AwardPrizeName", "The Prize Name field is required.");
-                }
+                ModelState.AddModelError("Trigger.AwardPrizeName", "The Prize Name field is required.");
             }
             if (model.AwardsMail)
             {
@@ -573,7 +606,11 @@ namespace GRA.Controllers.MissionControl
                         model.Trigger.LimitToSystemId = null;
                         model.Trigger.LimitToBranchId = null;
                         model.Trigger.LimitToProgramId = null;
-                        model.Trigger.SecretCode = model.Trigger.SecretCode.Trim().ToLower();
+                        model.Trigger.SecretCode = model
+                            .Trigger
+                            .SecretCode
+                            .Trim()
+                            .ToLowerInvariant();
                         model.Trigger.BadgeIds = new List<int>();
                         model.Trigger.ChallengeIds = new List<int>();
                     }
@@ -596,31 +633,30 @@ namespace GRA.Controllers.MissionControl
                     if (model.BadgeUploadImage != null
                         || !string.IsNullOrWhiteSpace(model.BadgeMakerImage))
                     {
-                        byte[] badgeBytes;
-                        string filename;
                         if (!string.IsNullOrWhiteSpace(model.BadgeMakerImage)
                             && (model.BadgeUploadImage == null || model.UseBadgeMaker))
                         {
                             var badgeString = model.BadgeMakerImage.Split(',').Last();
                             badgeBytes = Convert.FromBase64String(badgeString);
-                            filename = "badge.png";
                         }
                         else
                         {
-                            using (var fileStream = model.BadgeUploadImage.OpenReadStream())
+                            if (badgeBytes == null)
                             {
                                 using (var ms = new MemoryStream())
                                 {
-                                    fileStream.CopyTo(ms);
+                                    await model.BadgeUploadImage.CopyToAsync(ms);
                                     badgeBytes = ms.ToArray();
                                 }
                             }
-                            filename = Path.GetFileName(model.BadgeUploadImage.FileName);
                         }
+
                         var existing = await _badgeService
                                     .GetByIdAsync(model.Trigger.AwardBadgeId);
                         existing.Filename = Path.GetFileName(model.BadgePath);
-                        await _badgeService.ReplaceBadgeFileAsync(existing, badgeBytes);
+                        await _badgeService.ReplaceBadgeFileAsync(existing,
+                            badgeBytes,
+                            model.BadgeUploadImage.FileName);
                     }
                     var savedtrigger = await _triggerService.UpdateAsync(model.Trigger);
                     ShowAlertSuccess($"Trigger '<strong>{savedtrigger.Name}</strong>' was successfully modified");